+++ CHANGELOG.md
@@ -2,15 +2,19 @@
 
 BUG FIXES:
 
+* resource/functionality: Fixed bug that prevented the `description` attribute from being unset ([#49](https://github.com/firehydrant/terraform-provider-firehydrant/pull/49))
+* resource/functionality: Fixed bug that prevented the `services` attribute from being unset ([#49](https://github.com/firehydrant/terraform-provider-firehydrant/pull/49))
 * resource/service: Fixed bug that prevented the `description` attribute from being unset ([#51](https://github.com/firehydrant/terraform-provider-firehydrant/pull/51))
 * resource/service: Fixed bug that prevented `labels` from being removed from services ([#52](https://github.com/firehydrant/terraform-provider-firehydrant/pull/52))
 
 ENHANCEMENTS:
 
 * provider: Added Terraform version to the user agent header ([#24](https://github.com/firehydrant/terraform-provider-firehydrant/pull/24))
+* resource/functionality: Added deprecation warning to the `services` attribute, preferring `service_ids` instead ([#49](https://github.com/firehydrant/terraform-provider-firehydrant/pull/49))
 * resource/service: Added the `alert_on_add` attribute to services ([#24](https://github.com/firehydrant/terraform-provider-firehydrant/pull/24))
 * resource/service: Added the `owner_id` attribute to services ([#23](https://github.com/firehydrant/terraform-provider-firehydrant/pull/23))
 * resource/team: Added deprecation warning to the `services` attribute, preferring `service_ids` instead ([#31](https://github.com/firehydrant/terraform-provider-firehydrant/pull/31))
+* data_source/functionality: Added the `service_ids` attribute to functionality ([#49](https://github.com/firehydrant/terraform-provider-firehydrant/pull/49))
 * data_source/service: Added the `alert_on_add` attribute to services ([#24](https://github.com/firehydrant/terraform-provider-firehydrant/pull/24))
 * data_source/service: Added the `owner_id` attribute to services ([#23](https://github.com/firehydrant/terraform-provider-firehydrant/pull/23))
 * data_source/services: Added the `alert_on_add` attribute to services ([#24](https://github.com/firehydrant/terraform-provider-firehydrant/pull/24))
@@ -20,6 +24,8 @@ NOTES:
 
 * The deprecated attribute `services` will be removed from resource/team 3 months after the release of v0.2.0. You will have until May 31, 2022 to migrate to the preferred attribute.
    More information about this deprecation can be found in the description of ([#31](https://github.com/firehydrant/terraform-provider-firehydrant/pull/31))
+* The deprecated attribute `services` will be removed from resource/functionality 3 months after the release of v0.2.0. You will have until May 31, 2022 to migrate to the preferred attribute.
+  More information about this deprecation can be found in the description of ([#49](https://github.com/firehydrant/terraform-provider-firehydrant/pull/49))
 
 ## 0.1.4